exception handling - delte

+++ src/main/java/com/dvdthedev/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.dvdthedev.course.services;
 
 import com.dvdthedev.course.entities.User;
 import com.dvdthedev.course.repositories.UserRepository;
+import com.dvdthedev.course.services.exceptions.DatabaseException;
 import com.dvdthedev.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,14 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        try {
+            if (!userRepository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            } else
+                userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
